+++ internal/go_repository_tools_srcs.bzl
@@ -161,6 +161,9 @@ GO_REPOSITORY_TOOLS_SRCS = [
     Label("//v2/merger:merger.go"),
     Label("//v2/pathtools:BUILD.bazel"),
     Label("//v2/pathtools:path.go"),
+    Label("//v2/resolve:BUILD.bazel"),
+    Label("//v2/resolve:config.go"),
+    Label("//v2/resolve:index.go"),
     Label("//v2/rule:BUILD.bazel"),
     Label("//v2/rule:directives.go"),
     Label("//v2/rule:expr.go"),
