+++ resolve/index.go
@@ -16,23 +16,29 @@ limitations under the License.
 package resolve
 
 import (
-	"log"
+	"context"
 
+	"github.com/bazel-contrib/bazel-gazelle/v2/label"
+	v2 "github.com/bazel-contrib/bazel-gazelle/v2/resolve"
+	"github.com/bazel-contrib/bazel-gazelle/v2/rule"
 	"github.com/bazelbuild/bazel-gazelle/config"
-	"github.com/bazelbuild/bazel-gazelle/label"
 	"github.com/bazelbuild/bazel-gazelle/repo"
-	"github.com/bazelbuild/bazel-gazelle/rule"
 )
 
 // ImportSpec describes a library to be imported. Imp is an import string for
 // the library. Lang is the language in which the import string appears (this
 // should match Resolver.Name).
-type ImportSpec struct {
-	Lang, Imp string
-}
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.ImportSpec instead.
+//
+//go:fix inline
+type ImportSpec = v2.ImportSpec
 
 // Resolver is an interface that language extensions can implement to resolve
 // dependencies in rules they generate.
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.Indexer or
+// github.com/bazel-contrib/bazel-gazelle/v2/resolve.Resolver instead.
 type Resolver interface {
 	// Name returns the name of the language. This should be a prefix of the
 	// kinds of rules generated by the language, e.g., "go" for the Go extension
@@ -63,6 +69,8 @@ type Resolver interface {
 
 // CrossResolver is an interface that language extensions can implement to provide
 // custom dependency resolution logic for other languages.
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.CrossResolver instead.
 type CrossResolver interface {
 	// CrossResolve attempts to resolve an import string to a rule for languages
 	// other than the implementing extension. lang is the langauge of the rule
@@ -72,79 +80,74 @@ type CrossResolver interface {
 
 // RuleIndex is a table of rules in a workspace, indexed by label and by
 // import path. Used by Resolver to map import paths to labels.
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.RuleIndex instead.
 type RuleIndex struct {
-	mrslv          func(r *rule.Rule, pkgRel string) Resolver
-	crossResolvers []CrossResolver
-
-	// The underlying state of rules. All indexing should be reproducible from this.
-	rules []*ruleRecord
-
-	// If indexing of rules has occurred already
-	indexed bool
-
-	// Rules indexed by label.
-	// Computed from `rules` when indexing.
-	labelMap map[label.Label]*ruleRecord
-
-	// Imports specs mapping to records producing those.
-	// Computed from `rules` when indexing.
-	importMap map[ImportSpec][]*ruleRecord
-
-	// Whether another rule of the same language embeds this rule.
-	// Embedded rules should not be indexed.
-	// Computed from `rules` when indexing.
-	embedded map[label.Label]struct{}
-
-	// The transitive closure of labels embedded within rules (as determined by the Embeds method).
-	// This only includes rules in the same language (i.e., it includes a go_library embedding
-	// a go_proto_library, but not a go_proto_library embedding a proto_library).
-	// Computed from `rules` when indexing.
-	embeds map[label.Label][]label.Label
-
-	// The transitive closure of all imports produced by each label.
-	// This includes transitive imports from embedded labels (as determined by
-	// the Embeds method). This may include imports of other languages.
-	// Computed from `rules` when indexing.
-	imports map[label.Label][]ImportSpec
-}
-
-// ruleRecord contains information about a rule relevant to import indexing.
-type ruleRecord struct {
-	rule  *rule.Rule
-
-	Kind  string      `json:"kind"`
-	Label label.Label `json:"label"`
-
-	Pkg string `json:"pkg"`
-
-	// A list of ImportSpecs by which this rule may be imported.
-	ImportedAs []ImportSpec `json:"importedAs"`
-
-	// The set of labels (of any language) that this rule directly embeds.
-	Embeds []label.Label `json:"embeds"`
-
-	// The language that this rule is relevant for.
-	// Due to the presence of mapped kinds, it's otherwise
-	// impossible to know the underlying builtin rule type for an
-	// arbitrary import.
-	Lang string `json:"lang"`
+	v2 *v2.RuleIndex
 }
 
 // NewRuleIndex creates a new index.
 //
 // kindToResolver is a map from rule kinds (for example, "go_library") to
 // Resolvers that support those kinds.
-func NewRuleIndex(mrslv func(r *rule.Rule, pkgRel string) Resolver, exts ...interface{}) *RuleIndex {
-	var crossResolvers []CrossResolver
-	for _, e := range exts {
-		if cr, ok := e.(CrossResolver); ok {
-			crossResolvers = append(crossResolvers, cr)
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.NewRuleIndex instead.
+func NewRuleIndex(mrslv func(r *rule.Rule, pkgRel string) Resolver, exts ...any) *RuleIndex {
+	mrslvv2 := func(r *rule.Rule, pkgRel string) v2.Indexer {
+		rslv := mrslv(r, pkgRel)
+		if rslv == nil {
+			return nil
+		}
+		return resolverAdapter{v1: rslv}
+	}
+	var finders []v2.Finder
+	for _, ext := range exts {
+		if cr, ok := ext.(CrossResolver); ok {
+			finders = append(finders, crossResolverAdapter{v1: cr})
 		}
 	}
-	return &RuleIndex{
-		mrslv:          mrslv,
-		crossResolvers: crossResolvers,
-	}
+	indexv2 := v2.NewRuleIndex(mrslvv2, finders)
+	return WrapRuleIndexV2(indexv2)
+}
+
+func WrapRuleIndexV2(v2 *v2.RuleIndex) *RuleIndex {
+	return &RuleIndex{v2: v2}
+}
+
+type resolverAdapter struct {
+	v1 Resolver
+}
+
+var _ v2.Indexer = resolverAdapter{}
+var _ v2.Resolver = resolverAdapter{}
+
+func (a resolverAdapter) Name() string {
+	return a.v1.Name()
+}
+
+func (a resolverAdapter) Imports(ctx context.Context, args v2.ImportsArgs) (v2.ImportsResult, error) {
+	imps := a.v1.Imports(args.Config, args.Rule, args.File)
+	embeds := a.v1.Embeds(args.Rule, args.From)
+	return v2.ImportsResult{
+		Imports:       imps,
+		Embeds:        embeds,
+		NotImportable: imps == nil,
+	}, nil
+}
+
+func (a resolverAdapter) Resolve(ctx context.Context, args v2.ResolveArgs) error {
+	a.v1.Resolve(args.Config, WrapRuleIndexV2(args.Index), args.RemoteCache, args.Rule, args.Imports, args.From)
+	return nil
+}
+
+type crossResolverAdapter struct {
+	v1 CrossResolver
+}
+
+var _ v2.Finder = crossResolverAdapter{}
+
+func (a crossResolverAdapter) Find(ctx context.Context, args v2.FindArgs) ([]v2.FindResult, error) {
+	return a.v1.CrossResolve(args.Config, WrapRuleIndexV2(args.Index), args.Import, args.Lang), nil
 }
 
 // AddRule adds a rule r to the index. The rule will only be indexed if there
@@ -152,43 +155,10 @@ func NewRuleIndex(mrslv func(r *rule.Rule, pkgRel string) Resolver, exts ...inte
 // non-nil slice.
 //
 // AddRule may only be called before Finish.
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.RuleIndex.AddRule instead.
 func (ix *RuleIndex) AddRule(c *config.Config, r *rule.Rule, f *rule.File) {
-	if ix.indexed {
-		log.Fatal("AddRule called after Finish")
-	}
-
-	var lang string
-	var imps []ImportSpec
-	var embeds []label.Label
-
-	l := label.New(c.RepoName, f.Pkg, r.Name())
-
-	if rslv := ix.mrslv(r, f.Pkg); rslv != nil {
-		lang = rslv.Name()
-		if passesLanguageFilter(c.Langs, lang) {
-			imps = rslv.Imports(c, r, f)
-
-			for _, e := range rslv.Embeds(r, l) {
-				embeds = append(embeds, e.Abs(l.Repo, l.Pkg))
-			}
-		}
-	}
-	// If imps == nil, the rule is not importable. If imps is the empty slice,
-	// it may still be importable if it embeds importable libraries.
-	if imps == nil {
-		return
-	}
-
-	record := &ruleRecord{
-		rule:       r,
-		Kind:       r.Kind(),
-		Pkg:        f.Pkg,
-		Label:      l,
-		ImportedAs: imps,
-		Embeds:     embeds,
-		Lang:       lang,
-	}
-	ix.rules = append(ix.rules, record)
+	ix.v2.AddRule(c, r, f)
 }
 
 // Finish constructs the import index and performs any other necessary indexing
@@ -197,86 +167,16 @@ func (ix *RuleIndex) AddRule(c *config.Config, r *rule.Rule, f *rule.File) {
 //
 // Finish must be called after all AddRule calls and before any
 // FindRulesByImport calls.
+//
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.RuleIndex.Finish instead.
 func (ix *RuleIndex) Finish() {
-	ix.labelMap = make(map[label.Label]*ruleRecord)
-	ix.imports = make(map[label.Label][]ImportSpec)
-
-	for _, r := range ix.rules {
-		if _, ok := ix.labelMap[r.Label]; ok {
-			log.Printf("multiple rules found with label %s", r.Label)
-			continue
-		}
-
-		ix.labelMap[r.Label] = r
-		ix.imports[r.Label] = r.ImportedAs
-	}
-
-	ix.collectEmbeds()
-	ix.buildImportIndex()
-
-	ix.indexed = true
+	ix.v2.Finish()
 }
 
-func (ix *RuleIndex) collectEmbeds() {
-	ix.embeds = make(map[label.Label][]label.Label)
-	ix.embedded = make(map[label.Label]struct{})
-
-	didCollectEmbeds := make(map[label.Label]bool)
-
-	for _, r := range ix.rules {
-		ix.collectRecordEmbeds(r, didCollectEmbeds)
-	}
-}
-func (ix *RuleIndex) collectRecordEmbeds(r *ruleRecord, didCollectEmbeds map[label.Label]bool) {
-	if _, ok := didCollectEmbeds[r.Label]; ok {
-		return
-	}
-	resolver := ix.mrslv(r.rule, r.Pkg)
-	didCollectEmbeds[r.Label] = true
-	ix.embeds[r.Label] = r.Embeds
-	for _, e := range r.Embeds {
-		er, ok := ix.labelMap[e]
-		if !ok {
-			continue
-		}
-		ix.collectRecordEmbeds(er, didCollectEmbeds)
-		erResolver := ix.mrslv(er.rule, er.Pkg)
-		if resolver.Name() == erResolver.Name() {
-			ix.embedded[er.Label] = struct{}{}
-			ix.embeds[r.Label] = append(ix.embeds[r.Label], ix.embeds[er.Label]...)
-		}
-		ix.imports[r.Label] = append(ix.imports[r.Label], ix.imports[er.Label]...)
-	}
-}
-
-// buildImportIndex constructs the map used by FindRulesByImport.
-func (ix *RuleIndex) buildImportIndex() {
-	ix.importMap = make(map[ImportSpec][]*ruleRecord)
-	for _, r := range ix.rules {
-		if _, embedded := ix.embedded[r.Label]; embedded {
-			continue
-		}
-		indexed := make(map[ImportSpec]bool)
-		for _, imp := range ix.imports[r.Label] {
-			if indexed[imp] {
-				continue
-			}
-			indexed[imp] = true
-			ix.importMap[imp] = append(ix.importMap[imp], r)
-		}
-	}
-}
-
-type FindResult struct {
-	// Label is the absolute label (including repository and package name) for
-	// a matched rule.
-	Label label.Label
-
-	// Embeds is the transitive closure of labels for rules that the matched
-	// rule embeds. It may contains duplicates and does not include the label
-	// for the rule itself.
-	Embeds []label.Label
-}
+// Deprecated: Use github.com/bazel-contrib/bazel-gazelle/v2/resolve.FindResult instead.
+//
+//go:fix inline
+type FindResult = v2.FindResult
 
 // FindRulesByImport attempts to resolve an import string to a rule record.
 // imp is the import to resolve (which includes the target language). lang is
@@ -291,60 +191,12 @@ type FindResult struct {
 //
 // DEPRECATED: use FindRulesByImportWithConfig instead
 func (ix *RuleIndex) FindRulesByImport(imp ImportSpec, lang string) []FindResult {
-	matches := ix.importMap[imp]
-	results := make([]FindResult, 0, len(matches))
-	for _, m := range matches {
-		if m.Lang != lang {
-			continue
-		}
-		results = append(results, FindResult{
-			Label:  m.Label,
-			Embeds: ix.embeds[m.Label],
-		})
-	}
-	return results
+	return ix.v2.FindRulesByImport(imp, lang)
 }
 
 // FindRulesByImportWithConfig attempts to resolve an import to a rule first by
 // checking the rule index, then if no matches are found any registered
 // CrossResolve implementations are called.
 func (ix *RuleIndex) FindRulesByImportWithConfig(c *config.Config, imp ImportSpec, lang string) []FindResult {
-	results := ix.FindRulesByImport(imp, lang)
-	if len(results) > 0 {
-		return results
-	}
-	for _, cr := range ix.crossResolvers {
-		results = append(results, cr.CrossResolve(c, ix, imp, lang)...)
-	}
-	return results
-}
-
-// IsSelfImport returns true if the result's label matches the given label
-// or the result's rule transitively embeds the rule with the given label.
-// Self imports cause cyclic dependencies, so the caller may want to omit
-// the dependency or report an error.
-func (r FindResult) IsSelfImport(from label.Label) bool {
-	if from.Equal(r.Label) {
-		return true
-	}
-	for _, e := range r.Embeds {
-		if from.Equal(e) {
-			return true
-		}
-	}
-	return false
-}
-
-// passesLanguageFilter returns true if the filter is empty (disabled) or if the
-// given language name appears in it.
-func passesLanguageFilter(langFilter []string, langName string) bool {
-	if len(langFilter) == 0 {
-		return true
-	}
-	for _, l := range langFilter {
-		if l == langName {
-			return true
-		}
-	}
-	return false
+	return ix.v2.FindRulesByImportWithConfig(c, imp, lang)
 }
