+++ v2/BUILD.bazel
@@ -14,6 +14,7 @@ filegroup(
         "//v2/label:all_files",
         "//v2/merger:all_files",
         "//v2/pathtools:all_files",
+        "//v2/resolve:all_files",
         "//v2/rule:all_files",
         "//v2/testtools:all_files",
     ],
