+++ v2/resolve/index.go
@@ -1,0 +1,434 @@
+/* Copyright 2018 The Bazel Authors. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+   http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package resolve
+
+import (
+	"context"
+	"log"
+
+	"github.com/bazel-contrib/bazel-gazelle/v2/config"
+	"github.com/bazel-contrib/bazel-gazelle/v2/label"
+	"github.com/bazel-contrib/bazel-gazelle/v2/rule"
+	"github.com/bazelbuild/bazel-gazelle/repo"
+)
+
+// ImportSpec describes a library to be imported. Imp is an import string for
+// the library. Lang is the language in which the import string appears (this
+// should match Resolver.Name).
+type ImportSpec struct {
+	Lang, Imp string
+}
+
+// Indexer is an interface that language extensions can implement to list the
+// names by which a rule can be imported.
+type Indexer interface {
+	// TODO(v2): remove this
+	Name() string
+
+	// Imports returns the names by which a rule can be imported. These names are
+	// added to RuleIndex to aid dependency resolution.
+	Imports(context.Context, ImportsArgs) (ImportsResult, error)
+}
+
+type ImportsArgs struct {
+	// Config is the configuration directory that contains the rule.
+	Config *config.Config
+
+	// Rule is the non-nil rule to index.
+	Rule *rule.Rule
+
+	// File is the build file that contains Rule.
+	File *rule.File
+
+	// TODO(v2): try to refactor this out or document it
+	From label.Label
+}
+
+type ImportsResult struct {
+	// Imports is a list of language names and strings by which the rule may be
+	// imported.
+	Imports []ImportSpec
+
+	// Embeds is a list of library labels that this rule embeds. Both the embedder
+	// and embeddee may be importable by the same string, but in case of ambiguity,
+	// the dependency resolver prefers the embedder.
+	Embeds []label.Label
+
+	// NotImportable is set for a rule that can't be imported, such as a test.
+	// It's not necessary to set this when both Imports and Embeds are empty.
+	NotImportable bool
+}
+
+// Finder is an interface that language extensions can implement to provide
+// import resolution beyond what RuleIndex can do.
+//
+// Most extensions do not need to implement this interface. It may be useful
+// in cases where Indexer can't be completely implemented. For exmaple,
+// the proto extension's Imports method can return imports in each language
+// that it's aware of, but the user may have a custom proto compiler, rule set,
+// and Gazelle extension.
+type Finder interface {
+	// Find returns a list of libraries that can be imported with the given
+	// import string. Each extension's Find method is called by RuleIndex if
+	// a rule was not located within the index.
+	Find(context.Context, FindArgs) ([]FindResult, error)
+}
+
+type FindArgs struct {
+	// Config is the configuration for the current directory.
+	Config *config.Config
+
+	// Index is global rule index, built with [Indexer].
+	Index *RuleIndex
+
+	// Import is the import string to resolve. Find should return a list of
+	// libraries that may be imported with this string.
+	Import ImportSpec
+
+	// Lang is the language of the source code where Import was found.
+	Lang string
+}
+
+// Resolver is an interface that languages extensions can implement to transform
+// import strings from source files into Bazel labels, typically in a "deps"
+// attribute. This is done separately from GenerateRules so that Gazelle can
+// build an index of relevant library rules across the repo, including rules
+// generated by GenerateRules.
+type Resolver interface {
+	// Resolve performs dependency resolution on a rule, usually setting its
+	// "deps" attribute.
+	Resolve(context.Context, ResolveArgs) error
+}
+
+type ResolveArgs struct {
+	// Config is the configuration for the current directory.
+	Config *config.Config
+
+	// Index is a table mapping import strings to bazel labels for known
+	// libraries, constructed using Indexer.
+	Index *RuleIndex
+
+	// Rule is the rule to resolve dependencies for. Resolve should modify its
+	// "deps" attribute (or language-appropriate equivalent).
+	Rule *rule.Rule
+
+	// From is Rule's Bazel label.
+	From label.Label
+
+	// TODO(v2): definitely remove usage of this and refactor it out.
+	RemoteCache *repo.RemoteCache
+
+	// Imports is returned by GenerateRules.
+	// TODO(v2): try to remove all usage of this, then refactor it out.
+	Imports any
+}
+
+// RuleIndex is a table of rules in a workspace, indexed by label and by
+// import path. Used by Resolver to map import paths to labels.
+type RuleIndex struct {
+	mrslv   func(r *rule.Rule, pkgRel string) Indexer
+	finders []Finder
+
+	// The underlying state of rules. All indexing should be reproducible from this.
+	rules []*ruleRecord
+
+	// If indexing of rules has occurred already
+	indexed bool
+
+	// Rules indexed by label.
+	// Computed from `rules` when indexing.
+	labelMap map[label.Label]*ruleRecord
+
+	// Imports specs mapping to records producing those.
+	// Computed from `rules` when indexing.
+	importMap map[ImportSpec][]*ruleRecord
+
+	// Whether another rule of the same language embeds this rule.
+	// Embedded rules should not be indexed.
+	// Computed from `rules` when indexing.
+	embedded map[label.Label]struct{}
+
+	// The transitive closure of labels embedded within rules (as determined by the Embeds method).
+	// This only includes rules in the same language (i.e., it includes a go_library embedding
+	// a go_proto_library, but not a go_proto_library embedding a proto_library).
+	// Computed from `rules` when indexing.
+	embeds map[label.Label][]label.Label
+
+	// The transitive closure of all imports produced by each label.
+	// This includes transitive imports from embedded labels (as determined by
+	// the Embeds method). This may include imports of other languages.
+	// Computed from `rules` when indexing.
+	imports map[label.Label][]ImportSpec
+}
+
+// ruleRecord contains information about a rule relevant to import indexing.
+type ruleRecord struct {
+	rule *rule.Rule
+
+	Kind  string      `json:"kind"`
+	Label label.Label `json:"label"`
+
+	Pkg string `json:"pkg"`
+
+	// A list of ImportSpecs by which this rule may be imported.
+	ImportedAs []ImportSpec `json:"importedAs"`
+
+	// The set of labels (of any language) that this rule directly embeds.
+	Embeds []label.Label `json:"embeds"`
+
+	// The language that this rule is relevant for.
+	// Due to the presence of mapped kinds, it's otherwise
+	// impossible to know the underlying builtin rule type for an
+	// arbitrary import.
+	Lang string `json:"lang"`
+}
+
+// NewRuleIndex creates a new index.
+//
+// kindToResolver is a map from rule kinds (for example, "go_library") to
+// Resolvers that support those kinds.
+func NewRuleIndex(mrslv func(r *rule.Rule, pkgRel string) Indexer, finders []Finder) *RuleIndex {
+	return &RuleIndex{
+		mrslv:   mrslv,
+		finders: finders,
+	}
+}
+
+// AddRule adds a rule r to the index. The rule will only be indexed if there
+// is a known resolver for the rule's kind and Resolver.Imports returns a
+// non-nil slice.
+//
+// AddRule may only be called before Finish.
+func (ix *RuleIndex) AddRule(c *config.Config, r *rule.Rule, f *rule.File) {
+	if ix.indexed {
+		log.Fatal("AddRule called after Finish")
+	}
+
+	var lang string
+	var imps []ImportSpec
+	var embeds []label.Label
+
+	l := label.New(c.RepoName, f.Pkg, r.Name())
+
+	if rslv := ix.mrslv(r, f.Pkg); rslv != nil {
+		lang = rslv.Name()
+		if passesLanguageFilter(c.Langs, lang) {
+			result, err := rslv.Imports(context.TODO(), ImportsArgs{
+				Config: c,
+				Rule:   r,
+				File:   f,
+				From:   l,
+			})
+			if err != nil {
+				// TODO(v2): return
+				log.Print(err)
+				return
+			}
+			imps = result.Imports
+			for _, e := range result.Embeds {
+				embeds = append(embeds, e.Abs(l.Repo, l.Pkg))
+			}
+		}
+	}
+	// If imps == nil, the rule is not importable. If imps is the empty slice,
+	// it may still be importable if it embeds importable libraries.
+	if imps == nil {
+		return
+	}
+
+	record := &ruleRecord{
+		rule:       r,
+		Kind:       r.Kind(),
+		Pkg:        f.Pkg,
+		Label:      l,
+		ImportedAs: imps,
+		Embeds:     embeds,
+		Lang:       lang,
+	}
+	ix.rules = append(ix.rules, record)
+}
+
+// Finish constructs the import index and performs any other necessary indexing
+// actions after all rules have been added. This step is necessary because
+// a rule may be indexed differently based on what rules are added later.
+//
+// Finish must be called after all AddRule calls and before any
+// FindRulesByImport calls.
+func (ix *RuleIndex) Finish() {
+	ix.labelMap = make(map[label.Label]*ruleRecord)
+	ix.imports = make(map[label.Label][]ImportSpec)
+
+	for _, r := range ix.rules {
+		if _, ok := ix.labelMap[r.Label]; ok {
+			log.Printf("multiple rules found with label %s", r.Label)
+			continue
+		}
+
+		ix.labelMap[r.Label] = r
+		ix.imports[r.Label] = r.ImportedAs
+	}
+
+	ix.collectEmbeds()
+	ix.buildImportIndex()
+
+	ix.indexed = true
+}
+
+func (ix *RuleIndex) collectEmbeds() {
+	ix.embeds = make(map[label.Label][]label.Label)
+	ix.embedded = make(map[label.Label]struct{})
+
+	didCollectEmbeds := make(map[label.Label]bool)
+
+	for _, r := range ix.rules {
+		ix.collectRecordEmbeds(r, didCollectEmbeds)
+	}
+}
+func (ix *RuleIndex) collectRecordEmbeds(r *ruleRecord, didCollectEmbeds map[label.Label]bool) {
+	if _, ok := didCollectEmbeds[r.Label]; ok {
+		return
+	}
+	resolver := ix.mrslv(r.rule, r.Pkg)
+	didCollectEmbeds[r.Label] = true
+	ix.embeds[r.Label] = r.Embeds
+	for _, e := range r.Embeds {
+		er, ok := ix.labelMap[e]
+		if !ok {
+			continue
+		}
+		ix.collectRecordEmbeds(er, didCollectEmbeds)
+		erResolver := ix.mrslv(er.rule, er.Pkg)
+		if resolver.Name() == erResolver.Name() {
+			ix.embedded[er.Label] = struct{}{}
+			ix.embeds[r.Label] = append(ix.embeds[r.Label], ix.embeds[er.Label]...)
+		}
+		ix.imports[r.Label] = append(ix.imports[r.Label], ix.imports[er.Label]...)
+	}
+}
+
+// buildImportIndex constructs the map used by FindRulesByImport.
+func (ix *RuleIndex) buildImportIndex() {
+	ix.importMap = make(map[ImportSpec][]*ruleRecord)
+	for _, r := range ix.rules {
+		if _, embedded := ix.embedded[r.Label]; embedded {
+			continue
+		}
+		indexed := make(map[ImportSpec]bool)
+		for _, imp := range ix.imports[r.Label] {
+			if indexed[imp] {
+				continue
+			}
+			indexed[imp] = true
+			ix.importMap[imp] = append(ix.importMap[imp], r)
+		}
+	}
+}
+
+type FindResult struct {
+	// Label is the absolute label (including repository and package name) for
+	// a matched rule.
+	Label label.Label
+
+	// Embeds is the transitive closure of labels for rules that the matched
+	// rule embeds. It may contains duplicates and does not include the label
+	// for the rule itself.
+	Embeds []label.Label
+}
+
+// FindRulesByImport attempts to resolve an import string to a rule record.
+// imp is the import to resolve (which includes the target language). lang is
+// the language of the rule with the dependency (for example, in
+// go_proto_library, imp will have ProtoLang and lang will be GoLang).
+// from is the rule which is doing the dependency. This is used to check
+// vendoring visibility and to check for self-imports.
+//
+// FindRulesByImport returns a list of rules, since any number of rules may
+// provide the same import. Callers may need to resolve ambiguities using
+// language-specific heuristics.
+//
+// DEPRECATED: use FindRulesByImportWithConfig instead
+//
+// TODO(v2): Delete this.
+func (ix *RuleIndex) FindRulesByImport(imp ImportSpec, lang string) []FindResult {
+	matches := ix.importMap[imp]
+	results := make([]FindResult, 0, len(matches))
+	for _, m := range matches {
+		if m.Lang != lang {
+			continue
+		}
+		results = append(results, FindResult{
+			Label:  m.Label,
+			Embeds: ix.embeds[m.Label],
+		})
+	}
+	return results
+}
+
+// FindRulesByImportWithConfig attempts to resolve an import to a rule first by
+// checking the rule index, then if no matches are found any registered
+// CrossResolve implementations are called.
+//
+// TODO(v2): Rename to Find.
+func (ix *RuleIndex) FindRulesByImportWithConfig(c *config.Config, imp ImportSpec, lang string) []FindResult {
+	results := ix.FindRulesByImport(imp, lang)
+	if len(results) > 0 {
+		return results
+	}
+	for _, f := range ix.finders {
+		// TODO(v2): plumb context, handle error
+		rs, err := f.Find(context.TODO(), FindArgs{
+			Config: c,
+			Index:  ix,
+			Import: imp,
+			Lang:   lang,
+		})
+		if err != nil {
+			log.Print(err)
+		}
+		results = append(results, rs...)
+	}
+	return results
+}
+
+// IsSelfImport returns true if the result's label matches the given label
+// or the result's rule transitively embeds the rule with the given label.
+// Self imports cause cyclic dependencies, so the caller may want to omit
+// the dependency or report an error.
+func (r FindResult) IsSelfImport(from label.Label) bool {
+	if from.Equal(r.Label) {
+		return true
+	}
+	for _, e := range r.Embeds {
+		if from.Equal(e) {
+			return true
+		}
+	}
+	return false
+}
+
+// passesLanguageFilter returns true if the filter is empty (disabled) or if the
+// given language name appears in it.
+func passesLanguageFilter(langFilter []string, langName string) bool {
+	if len(langFilter) == 0 {
+		return true
+	}
+	for _, l := range langFilter {
+		if l == langName {
+			return true
+		}
+	}
+	return false
+}
